// src/forgetrustee/frontend/src/types/generated/metadataModuleDto.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { MetadataHintDto } from './metadataHintDto';
import type { MetadataModuleDtoFields } from './metadataModuleDtoFields';
import type { MetadataModuleDtoSections } from './metadataModuleDtoSections';

export interface MetadataModuleDto {
  moduleId?: string;
  requestDtoClassName?: string;
  order?: number;
  label?: string;
  heading?: string;
  menuPoint?: string;
  tooltip?: string;
  helpText?: string;
  callToAction?: string;
  examples?: string[];
  hints?: MetadataHintDto[];
  fields?: MetadataModuleDtoFields;
  sections?: MetadataModuleDtoSections;
}